requirements.txt
flask~=3.0.0
splitwise~=3.0.0
python-dotenv~=1.0.0
pandas~=2.1.1
plotly~=5.17.0
python-dateutil~=2.8.2